requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Firebase
firebase-admin==6.0.0
google-cloud-firestore==2.13.1

# Machine Learning & Data Processing
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0
joblib==1.3.2

# Data processing utilities
python-dateutil==2.8.2
pytz==2023.3

# API utilities
httpx==0.25.2
python-multipart==0.0.6